chore: rename files for consistency
